# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API CineVault
  description: Documentación oficial de la API de CineVault (red social de cine).
  version: 1.0.0
  contact:
    name: Equipo CineVault
servers:
  - url: https://api.cinevault.art/api
    description: Producción
  - url: http://localhost:4000/api
    description: Desarrollo local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
...
    Resena:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        movie_id: { type: integer }
        content: { type: string }
        rating: { type: number, minimum: 1, maximum: 5 }
        likes: { type: integer }
        created_at: { type: string, format: date-time }
    Comentario:
      type: object
      properties:
        id: { type: integer }
        review_id: { type: integer }
        user_id: { type: integer }
        content: { type: string }
        created_at: { type: string, format: date-time }
...
    CreateReviewRequest:
      type: object
      required: [movie_id, rating]
      properties:
        movie_id: { type: integer }
        rating: { type: number, minimum: 1, maximum: 5, multipleOf: 0.5 }
        content: { type: string }
    UpdateReviewRequest:
      type: object
      properties:
        rating: { type: number, minimum: 1, maximum: 5, multipleOf: 0.5 }
        content: { type: string }
    ReportReviewRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
...
paths:
...
  /reviews:
    get:
      tags: [Reseñas]
      summary: Mis reseñas
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Resena' } } } } }
    post:
      tags: [Reseñas]
      summary: Crear reseña
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201': { description: Creada }
  /reviews/user/{userId}:
    get:
      tags: [Reseñas]
      summary: Reseñas de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista }
  /reviews/movie/{movieId}:
    get:
      tags: [Reseñas]
      summary: Reseñas por película
      parameters:
        - in: path
          name: movieId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista }
  /reviews/{reviewId}:
    patch:
      tags: [Reseñas]
      summary: Actualizar reseña
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200': { description: Actualizada }
    delete:
      tags: [Reseñas]
      summary: Eliminar reseña
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Eliminada }
  /reviews/{reviewId}/like:
    post:
      tags: [Reseñas]
      summary: Dar like
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      responses:
        '201': { description: Like añadido }
    delete:
      tags: [Reseñas]
      summary: Quitar like
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Like eliminado }
  /reviews/{reviewId}/report:
    post:
      tags: [Reseñas]
      summary: Reportar reseña
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportReviewRequest'
      responses:
        '200': { description: Reporte creado }
  /reviews/{reviewId}/comments:
    get:
      tags: [Comentarios]
      summary: Listar comentarios
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Comentario' } } } } }
    post:
      tags: [Comentarios]
      summary: Crear comentario
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: reviewId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201': { description: Creado }
...